+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## [1.6.0] (July 20, 2021)
 
 FEATURES: